Feat: implemented ideas from trello in summary
