+++ src/models/HistoryData.ts
@@ -1,0 +1,7 @@
+type HistoryData = {
+  date: number;
+  description: string;
+  amount: string | number;
+};
+
+export default HistoryData;
